Refactor: 서비스 로직을 각 도메인에 알맞은 디렉토리로 이동
