+++ src/main/java/com/team7/retriever/domain/channel/controller/ChannelCheckController.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.domain.channel.controller;
 
 import com.team7.retriever.domain.channel.controller.dto.response.ChannelInfoResponse;
-import com.team7.retriever.service.ChannelInfoService;
+import com.team7.retriever.domain.channel.service.ChannelInfoService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
