+++ src/main/java/com/team7/retriever/domain/channel/controller/ChannelInfoController.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.domain.channel.controller;
 
 import com.team7.retriever.domain.channel.controller.dto.request.ChannelInfoRequest;
-import com.team7.retriever.service.ChannelInfoService;
+import com.team7.retriever.domain.channel.service.ChannelInfoService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
