+++ src/main/java/com/team7/retriever/domain/crawling/service/PreprocessService.java
@@ -1,11 +1,14 @@
-package com.team7.retriever.service;
+package com.team7.retriever.domain.crawling.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.team7.retriever.domain.channel.service.ChInfoService;
+import com.team7.retriever.domain.channel.service.ChannelInfoService;
 import com.team7.retriever.domain.crawling.controller.dto.response.PreprocessResponse;
 import com.team7.retriever.domain.post.domain.document.Posts;
 import com.team7.retriever.domain.post.domain.repository.PostsRepository;
+import com.team7.retriever.domain.post.service.PostsService;
 
 import lombok.RequiredArgsConstructor;
