+++ src/main/java/com/team7/retriever/domain/drug/controller/DrugsController.java
@@ -2,7 +2,7 @@ package com.team7.retriever.domain.drug.controller;
 
 
 import com.team7.retriever.domain.drug.domain.document.Drugs;
-import com.team7.retriever.service.DrugsService;
+import com.team7.retriever.domain.drug.service.DrugsService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
