+++ src/main/java/com/team7/retriever/domain/post/controller/PostUpdateCheckController.java
@@ -2,9 +2,9 @@ package com.team7.retriever.domain.post.controller;
 
 import com.team7.retriever.domain.post.controller.dto.request.PostUpdateRequest;
 import com.team7.retriever.domain.post.controller.dto.request.PostUpdateCheckRequest;
-import com.team7.retriever.service.PostsService;
-import com.team7.retriever.service.PreprocessService;
-import com.team7.retriever.service.UpdateCheckService;
+import com.team7.retriever.domain.post.service.PostsService;
+import com.team7.retriever.domain.crawling.service.PreprocessService;
+import com.team7.retriever.domain.post.service.PostUpdateCheckService;
 import lombok.AllArgsConstructor;
 
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +17,7 @@ import java.util.List;
 public class PostUpdateCheckController {
     private final PostsService postsService;
     private final PreprocessService preprocessService;
-    private final UpdateCheckService updateCheckService;
+    private final PostUpdateCheckService postUpdateCheckService;
 
     // DB 데이터 조회 결과 테스트
     @GetMapping("/posts")
@@ -27,7 +27,7 @@ public class PostUpdateCheckController {
 
     @GetMapping
     public void updateAllPostsTest() {
-        updateCheckService.updateAllPost();
+        postUpdateCheckService.updateAllPost();
     }
 
     // Post 업데이트 테스트
