+++ src/main/java/com/team7/retriever/domain/post/controller/PostsController.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.domain.post.controller;
 
 import com.team7.retriever.domain.post.domain.document.Posts;
-import com.team7.retriever.service.PostsService;
+import com.team7.retriever.domain.post.service.PostsService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
