+++ src/main/java/com/team7/retriever/domain/post/service/PostSimilarityService.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.service;
+package com.team7.retriever.domain.post.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
